+++ CHANGES.md
@@ -9,13 +9,15 @@
 - FIX
   - バグ修正
 
-## 2025.4
+## 2026.1
 
 - [CHANGE] サイマルキャストの設定値である SimulcastRid を SimulcastRequestRid に移行する
   - これによりシステム条件の WebRTC SFU Sora のバージョンが 2025.2.0 以降になる
   - @zztkm
 - [UPDATE] システム条件を更新する
   - WebRTC SFU Sora 2025.2.0 以降
+  - @zztkm
+- [UPDATE] Sora Android SDK を 2026.1.0 に上げる
   - @zztkm
 
 ### misc